Make PDF Binary be saved in memory to prevent redundant fetching
